backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
celery[redis]==5.4.0
redis==5.1.0
sqlalchemy==2.0.36
alembic==1.13.3
psycopg2-binary==2.9.10
python-multipart==0.0.12
boto3==1.35.50
stripe==11.1.0
pydantic-settings==2.6.0
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
